finishing what i started (and cake people!))
hopefully this fixes some bugs that were made present due to a botched PR I did a few days ago, as well as properly implementing craftable cake people (or cake golems, if you wanna call them that). sadly the code that was present when i tried to make my modifications for some reason does not give the person attempting to eat the cake person any nutriment or vitamins, so i would appreciate it if a more experienced coder gave me a hint on how to fix this, or just fixed it themselves.