Calorite golems!
Calorite golems are a new type of golem that can be made using 10 sheets of calorite that, thanks to the combined magical forces of both adamantine and calorite, can use magic to heal the wounded, make food more nourishing, and make people around them faster for a moment. Unfortunately, calorite golems lack the armor that other golems have, and their fists cannot stun or do much damage on their own.