Fixes and issues pt. 1
so i forgot to add the vendor and loadout stuff for the outfits, but in trying to fix it, realized that updating the vendors would be good too; eeexcept that in the vendor lists are a buncha items that GS doesn't have. they're not in this commit, but i'm putting this forward now because i'm done working on this for the night.