Organizing our \code\modules, minor fixes, changes, and additions.
Moved the Gato burrito to under food_and_drinks.
Added a recipe for Flatulose under fatchem.dm.
Fixed some emote visibility and formatting (need to figure out how to remove the space in the *gurgle emote still).
Buffed the fizulphite and extilphite production in it's recipe to be in-line with the volume of it's ingredients, rather than producing less than the sum of it's parts.