Fatness Hiding System and Bluespace Belt
Applied changes to fatness.dm to allow the creation of objects, events and such that hide a character's actual fatness level to display and treat it as another.
Fixes to other files to support said change.
Added bluespace belt, an item that hides a character's current weight and treats it as if were 0. (Belt has currently no way to be created/accessed besides admin spawns. Haven't tested ALL things that interact with weight yet)