Halt autoclose when blocked too many times
Wooden barricades in powered airlocks cause lag as the doors try to autoclose over and over.

This change adds a failure count that increments each time a door attempts to close but something is blocking it. If it fails too many times, it simply stops trying. Doors can be closed manually once the blockage is removed.