Add additional weight to taur characters
Adds additional weight to taur characters when they go on scales. It's 2.5x for taur characters, meaning at 0 fatness a taur character will report a weight of 350 pounds while a non-taur character will report a weight of 140 pounds.
Math was based on the assumption of a taur body adding at least 200 or so pounds, then making it multiplicative to imply the tauric half was also gaining weight.