Barrel Tank transfer variable
Added transfer_amount to the barrel tank to dynamically change the amount of of chemicals transferred from the tank to the character (this is then multiplied by 2, as per the standard barrel tank function).
This variable is DIFFERENT from amount_per_transfer_from_this, as that is part of the reagent containers code and is set to 0 from the barrel tank by default.