Dumbbell and treadmill
Added dumbbell, a item held in hand that can be used to burn some fat. Can be used faster the more it is used consecutively.
Added treadmill, a placed item that functions as a conveyor belt (activated with a switch too), that burns some fat of people moved by it